LayoutTests/imported/w3c/web-platform-tests/intersection-observer/idlharness.window-expected.txt

CONSOLE MESSAGE: line 8: ReferenceError: Can't find variable: idl_test

FAIL Untitled ReferenceError: Can't find variable: idl_test
